fog-file-insights/django_cloud/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
django-storages
boto3
